moved model since cannot find module on github action
