+++ auth/src/routes/signin.ts
@@ -1,7 +1,7 @@
 import express, { Request, Response } from "express";
 import { body } from "express-validator";
 import { validateRequest, BadRequestError } from "@ksticketing/common"
-import { User } from "../models/user";
+import { User } from "../../models/user";
 import { Password } from "../services/password";
 import jwt from "jsonwebtoken";
